youtube-rag-extension/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
youtube-transcript-api
langchain
langchain-google-genai
google-generativeai
faiss-cpu
requests
pydantic